+++ feud/_internal/_decorators.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
@@ -15,6 +15,8 @@ import pydantic_core as pydc
 
 from feud import click
 
+AnyCallableT = t.TypeVar("AnyCallableT", bound=t.Callable[..., t.Any])
+
 
 def validate_call(
     func: t.Callable,
@@ -27,7 +29,7 @@ def validate_call(
     positional: list[str],
     var_positional: str | None,
     pydantic_kwargs: dict[str, t.Any],
-) -> t.Callable:
+) -> t.Callable[[AnyCallableT], AnyCallableT]:
     @ft.wraps(func)
     def wrapper(*args: t.Any, **kwargs: t.Any) -> t.Callable:
         try:
@@ -38,22 +40,31 @@ def validate_call(
                     args += (pos_arg,)
 
             # move *args to positional arguments
-            var_pos_args = kwargs.pop(
-                var_positional,
-                inspect._empty,  # noqa: SLF001
-            )
-            if var_pos_args is not inspect._empty:  # noqa: SLF001
-                args += var_pos_args
+            if var_positional is not None:
+                var_pos_args = kwargs.pop(
+                    var_positional,
+                    inspect._empty,  # noqa: SLF001
+                )
+                if var_pos_args is not inspect._empty:  # noqa: SLF001
+                    args += var_pos_args
 
             # apply renaming for any options
             inv_mapping = {v: k for k, v in param_renames.items()}
             true_kwargs = {inv_mapping.get(k, k): v for k, v in kwargs.items()}
 
             # create Pydantic configuration
-            config = pyd.ConfigDict(**pydantic_kwargs)
+            config = pyd.ConfigDict(
+                **pydantic_kwargs,  # type: ignore[typeddict-item]
+            )
 
             # validate the function call
-            return pyd.validate_call(func, config=config)(*args, **true_kwargs)
+            return pyd.validate_call(  # type: ignore[call-overload]
+                func,
+                config=config,
+            )(
+                *args,
+                **true_kwargs,
+            )
         except pyd.ValidationError as e:
             msg = re.sub(
                 r"validation error(s?) for (.*)\n",
@@ -85,4 +96,4 @@ def validate_call(
             )
             raise click.ClickException(msg) from None
 
-    return wrapper
+    return wrapper  # type: ignore[return-value]