+++ feud/_internal/_group.py
@@ -1,14 +1,21 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
 
+from __future__ import annotations
+
+import typing as t
+
 from feud import click
-from feud._internal import _command
+from feud._internal import _command, _meta
+
+if t.TYPE_CHECKING:
+    from feud.core.group import Group
 
 
-def get_group(__cls: type, /) -> click.Group:  # type[Group]
-    func: callable = __cls.__main__
+def get_group(__cls: type[Group], /) -> click.Group:  # type[Group]
+    func: t.Callable = __cls.__main__
     if isinstance(func, staticmethod):
         func = func.__func__
 
@@ -16,13 +23,7 @@ def get_group(__cls: type, /) -> click.Group:  # type[Group]
         config=__cls.__feud_config__,
         click_kwargs=__cls.__feud_click_kwargs__,
         is_group=True,
-        aliases=getattr(func, "__feud_aliases__", {}),
-        envs=getattr(func, "__feud_envs__", {}),
-        names=getattr(
-            func,
-            "__feud_names__",
-            _command.NameDict(command=None, params={}),
-        ),
+        meta=getattr(func, "__feud__", _meta.FeudMeta()),
         overrides={
             override.name: override
             for override in getattr(func, "__click_params__", [])
@@ -35,7 +36,7 @@ def get_group(__cls: type, /) -> click.Group:  # type[Group]
     )
 
     # generate click.Group and attach original function reference
-    command = state.decorate(func)
-    command.__func__ = func
-    command.__group__ = __cls
+    command: click.Group = state.decorate(func)  # type: ignore[assignment]
+    command.__func__ = func  # type: ignore[attr-defined]
+    command.__group__ = __cls  # type: ignore[attr-defined]
     return command