+++ feud/_internal/_sections.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
@@ -7,28 +7,18 @@ from __future__ import annotations
 
 import dataclasses
 from collections import defaultdict
-from typing import TypedDict
 
 from feud import click
+from feud._internal import _meta
 
 
-class CommandGroup(TypedDict):
-    name: str
-    commands: list[str]
+def add_command_sections(group: click.Group, context: list[str]) -> None:
+    from rich_click.utils import CommandGroupDict
 
-
-class OptionGroup(TypedDict):
-    name: str
-    options: list[str]
-
-
-def add_command_sections(
-    group: click.Group, context: list[str]
-) -> click.Group:
     if feud_group := getattr(group, "__group__", None):
-        command_groups: dict[str, list[CommandGroup]] = {
+        command_groups: dict[str, list[CommandGroupDict]] = {
             " ".join(context): [
-                CommandGroup(
+                CommandGroupDict(
                     name=section.name,
                     commands=[
                         item if isinstance(item, str) else item.name()
@@ -40,7 +30,7 @@ def add_command_sections(
         }
 
         for sub in group.commands.values():
-            if isinstance(sub, click.Group):
+            if sub.name and isinstance(sub, click.Group):
                 add_command_sections(sub, context=[*context, sub.name])
 
         settings = group.context_settings
@@ -56,10 +46,12 @@ def add_command_sections(
 
 def add_option_sections(
     obj: click.Command | click.Group, context: list[str]
-) -> click.Command | click.Group:
+) -> None:
     if isinstance(obj, click.Group):
         update_command(obj, context=context)
         for sub in obj.commands.values():
+            if sub.name is None:
+                continue
             if isinstance(sub, click.Group):
                 add_option_sections(sub, context=[*context, sub.name])
             else:
@@ -69,8 +61,11 @@ def add_option_sections(
 
 
 def get_opts(option: str, *, command: click.Command) -> list[str]:
+    func = command.__func__  # type: ignore[attr-defined]
     name_map = lambda name: name  # noqa: E731
-    if names := getattr(command.__func__, "__feud_names__", None):
+    meta: _meta.FeudMeta | None = getattr(func, "__feud__", None)
+    if meta and meta.names:
+        names = meta.names
         name_map = lambda name: names["params"].get(name, name)  # noqa: E731
     return next(
         param.opts
@@ -80,15 +75,19 @@ def get_opts(option: str, *, command: click.Command) -> list[str]:
 
 
 def update_command(command: click.Command, context: list[str]) -> None:
-    if func := getattr(command, "__func__", None):  # noqa: SIM102
-        if options := getattr(func, "__feud_sections__", None):
+    from rich_click.utils import OptionGroupDict
+
+    if func := getattr(command, "__func__", None):
+        meta: _meta.FeudMeta | None = getattr(func, "__feud__", None)
+        if meta and meta.sections:
+            options = meta.sections
             sections = defaultdict(list)
             for option, section_name in options.items():
                 opts: list[str] = get_opts(option, command=command)
                 sections[section_name].append(opts[0])
-            option_groups: dict[str, list[OptionGroup]] = {
+            option_groups: dict[str, list[OptionGroupDict]] = {
                 " ".join(context): [
-                    OptionGroup(name=name, options=options)
+                    OptionGroupDict(name=name, options=options)
                     for name, options in sections.items()
                 ]
             }