+++ feud/click/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
@@ -12,8 +12,8 @@ try:
     from rich_click import *
 
     is_rich = True
-except ImportError:
-    from click import *
+except (ImportError, ModuleNotFoundError):
+    from click import *  # type: ignore[no-redef, assignment]
 
     is_rich = False