+++ feud/core/command.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
@@ -120,4 +120,4 @@ def command(
             __func, config=cfg, click_kwargs=click_kwargs
         )
 
-    return decorate(func) if func else decorate
+    return decorate(func) if func else decorate  # type: ignore[return-value]