+++ feud/decorators.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
@@ -13,7 +13,7 @@ import typing as t
 
 import pydantic as pyd
 
-from feud._internal import _command
+from feud._internal import _meta
 from feud.exceptions import CompilationError
 
 __all__ = ["alias", "env", "rename", "section"]
@@ -123,9 +123,17 @@ def alias(**aliases: str | list[str]) -> t.Callable:
             )
             raise CompilationError(msg)
 
-        f.__feud_aliases__ = {
+        aliases_ = {
             k: ([v] if isinstance(v, str) else v) for k, v in aliases.items()
         }
+
+        if meta := getattr(f, "__feud__", None):
+            meta.aliases = aliases_
+        else:
+            f.__feud__ = _meta.FeudMeta(  # type: ignore[attr-defined]
+                aliases=aliases_,
+            )
+
         return f
 
     return decorator
@@ -193,7 +201,13 @@ def env(**envs: str) -> t.Callable:
             )
             raise CompilationError(msg)
 
-        f.__feud_envs__ = envs
+        if meta := getattr(f, "__feud__", None):
+            meta.envs = envs
+        else:
+            f.__feud__ = _meta.FeudMeta(  # type: ignore[attr-defined]
+                envs=envs,
+            )
+
         return f
 
     return decorator
@@ -255,7 +269,15 @@ def rename(command: str | None = None, /, **params: str) -> t.Callable:
             )
             raise CompilationError(msg)
 
-        f.__feud_names__ = _command.NameDict(command=command, params=params)
+        names = _meta.NameDict(command=command, params=params)
+
+        if meta := getattr(f, "__feud__", None):
+            meta.names = names
+        else:
+            f.__feud__ = _meta.FeudMeta(  # type: ignore[attr-defined]
+                names=names,
+            )
+
         return f
 
     return decorator
@@ -305,7 +327,15 @@ def section(**options: str) -> t.Callable:
             )
             raise CompilationError(msg)
 
-        f.__feud_sections__ = options.copy()
+        sections = options.copy()
+
+        if meta := getattr(f, "__feud__", None):
+            meta.sections = sections
+        else:
+            f.__feud__ = _meta.FeudMeta(  # type: ignore[attr-defined]
+                sections=sections,
+            )
+
         return f
 
     return decorator