+++ make/__init__.py
@@ -1,8 +1,8 @@
-# Copyright (c) 2023-2025 Feud Developers.
+# Copyright (c) 2023 Feud Developers.
 # Distributed under the terms of the MIT License (see the LICENSE file).
 # SPDX-License-Identifier: MIT
 # This source code is part of the Feud project (https://feud.wiki).
 
 """Tasks for package development."""
 
-from . import cov, docs, lint, release, tests
+from . import cov, docs, lint, release, tests, types